requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
colorama>=0.4.4